Decompile RIC EntityEntFactory
