+++ config/splat.us.ric.yaml
@@ -46,7 +46,8 @@ segments:
       - [0x1A800, .rodata, 20920]
       - [0x1A80C, .rodata, 21250]
       - [0x1A904, .rodata, 22380]
-      - [0x1A918, rodata]
+      - [0x1A918, .rodata, 24788]
+      - [0x1A968, rodata]
       - [0x1A9D0, .rodata, 26C84]
       - [0x1A9E8, rodata]
       - [0x1AB04, rodata]
@@ -61,6 +62,7 @@ segments:
       - [0x20920, c, 20920]
       - [0x21250, c, 21250]
       - [0x22380, c, 22380]
+      - [0x24788, c, 24788]
       - [0x26C84, c, 26C84]
       - [0x32324, c, 32324]
       - [0x345EC, c, 345EC]
